+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+* Issue **#28** : Added `encodeUrl` and `decodeUrl` functions.
+
+* Issue **#29** : The `link` function now encodes title, url and type to ensure these values don't break the link format.
+
 ## [v2.0.6] - 2019-01-04
 
 * Migrated to `java-library` plugin.
@@ -150,7 +154,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 * Initial release
 
-[Unreleased]: https://github.com/gchq/stroom-expression/compare/v1.4.14...HEAD
+[Unreleased]: https://github.com/gchq/stroom-expression/compare/v1.4.15...HEAD
+[v1.4.15]: https://github.com/gchq/stroom-expression/compare/v1.4.14...v1.4.15
 [v1.4.14]: https://github.com/gchq/stroom-expression/compare/v1.4.13...v1.4.14
 [v1.4.13]: https://github.com/gchq/stroom-expression/compare/v1.4.12...v1.4.13
 [v1.4.12]: https://github.com/gchq/stroom-expression/compare/v1.4.11...v1.4.12